+++ library/src/schemas/nullable/nullable.ts
@@ -11,7 +11,7 @@ export type NullableSchema<
     | Input<TWrapped>
     | (() => Input<TWrapped> | undefined)
     | undefined = undefined,
-  TOutput = TDefault extends Input<TWrapped>
+  TOutput = TDefault extends Input<TWrapped> | (() => Input<TWrapped>)
     ? Output<TWrapped>
     : Output<TWrapped> | null
 > = BaseSchema<Input<TWrapped> | null, TOutput> & {